chromium/third_party/blink/web_tests/webcodecs/videoframe-imagebitmap.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
  'use strict';
  function testCanvas_0f0(ctx, width, height, assert_compares) {
    let colorData = ctx.getImageData(0, 0, width, height).data;
    for (let i = 0; i < width * height; i += 4) {
      assert_compares(colorData[0], 0);
      assert_compares(colorData[1], 255);
      assert_compares(colorData[2], 0);
      assert_compares(colorData[3], 255);
    }
  }

  function testImageBitmapToAndFromVideoFrame(width, height) {
    let canvas = document.createElement('canvas');
    canvas.width = width;
    canvas.height = height;
    let ctx = canvas.getContext('2d');
    ctx.fillStyle = '#0f0';
    ctx.fillRect(0, 0, width, height);
    testCanvas_0f0(ctx, width, height, assert_equals);

    return createImageBitmap(canvas)
    .then((fromImageBitmap) => {
      let videoFrame = new VideoFrame({ timestamp: 0 }, fromImageBitmap);
      return videoFrame.createImageBitmap();
    })
    .then((toImageBitmap) => {
        let myCanvas = document.createElement('canvas');
        myCanvas.width = width;
        myCanvas.height = height;
        let myCtx = myCanvas.getContext('2d');
        myCtx.drawImage(toImageBitmap, 0, 0);
        let tolerance = 2;
        testCanvas_0f0(myCtx, width, height, (actual, expected) => {
          assert_approx_equals(actual, expected, tolerance);
        });
    });
  }

  promise_test(() => {
    return testImageBitmapToAndFromVideoFrame(320, 240);
  }, 'ImageBitmap<->VideoFrame with canvas size 320x240.');

  promise_test(() => {
    return testImageBitmapToAndFromVideoFrame(640, 480);
  }, 'ImageBitmap<->VideoFrame with canvas size 640x480.');

</script>
